feat: remove redundant, old, or unused code
