+++ src/app/(home)/@leaderboardTable/loading.tsx
@@ -1,0 +1,2 @@
+import { LeadearboardLoading } from "@/app/_shared/@leadearboard";
+export default LeadearboardLoading;
