+++ src/app/(home)/@leaderboardTable/page.tsx
@@ -1,0 +1,2 @@
+import { LeadearboardPage } from "@/app/_shared/@leadearboard";
+export default LeadearboardPage;
