--- src/shared/interfaces/index.jsx
@@ -1,5 +1,0 @@
-import { LeaderboardUser } from "./leaderboard-user";
-import { User } from "./user";
-import { UserStats } from "./user-stats";
-import { LeaderboardData } from "./leaderboard-data";
-export { LeaderboardUser, User, UserStats, LeaderboardData };
