--- src/shared/interfaces/leaderboard-data.tsx
@@ -1,5 +1,0 @@
-export interface LeaderboardData {
-  id: number;
-  user_id: number;
-  rank: number;
-}
