--- src/shared/interfaces/user-stats.tsx
@@ -1,6 +1,0 @@
-export interface UserStats {
-  user_id: number;
-  boss_score: number;
-  builder_score: number;
-  nominations: number;
-}
