--- src/shared/interfaces/user.tsx
@@ -1,5 +1,0 @@
-export interface User {
-  id: number;
-  wallet_address: string;
-  username: string;
-}
